Create README.md
##Hello there! 
I am a musician ( guitar player ), entrepreneur, full stack developer (in construction). My passion is to create valuable things for the community, that's the why of my learning.

Superjonic (superjonico in spanish): it's a concept created by me, it's about compose music overlaping differents ionan scales, the tones and sounds as a result are very interesting.